requirements.txt
djangorestframework
PyMySQL==0.7.11
Django==2.0
